+++ src/hooks/react-query/useSurvey.ts
@@ -1,5 +1,7 @@
 import {saveHealthSurvey} from '../../api/Survey';
-import {useMutation} from '@tanstack/react-query';
+import {useMutation, useQuery} from '@tanstack/react-query';
+import {fetchDiseaseList} from '../../api/MyPage';
+import {DiseaseItem} from '../../types/server/surveyList';
 
 export const useSaveHealthSurvey = () => {
     const { mutate } = useMutation({
@@ -7,4 +9,15 @@ export const useSaveHealthSurvey = () => {
     });
     
     return { mutate };
+};
+
+export const useQuery_DiseaseList = () => {
+    return useQuery({
+        queryKey: ['diseaseList'],
+        queryFn: fetchDiseaseList,
+        select: (response): DiseaseItem[] => response.data.data.map((item: { dateTime: string; }) => ({
+            ...item,
+            dateTime: item.dateTime.split('T')[0],
+        })),
+    });
 };
